% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateExtensionCandidates.R
\name{GenerateExtensionCandidates}
\alias{GenerateExtensionCandidates}
\title{GenerateExtensionCandidates}
\usage{
GenerateExtensionCandidates(bedtools_loc = NULL)
}
\value{
Rank ordered list of gene extension candidates saved to working directory as gene_extension_candidates.csv
}
\description{
Identifies candidate genes for extension with excess 3' intergenic reads
and create a rank ordered list of genes as a function of 3' intergenic read
mapping within 10kb of known gene end. You can use this as a prioritized gene
list for gene extension to examine in Integrated Genomics Viewer.

Note: It runs partially in bash/linux terminal. Make sure bedtools is installed and
in PATH variable in Linux or MacOS.
}
\examples{
GenerateExtensionCandidates(bedtools_loc)
}
