% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateGeneLocationBed.R
\name{GenerateGeneLocationBed}
\alias{GenerateGeneLocationBed}
\title{GenerateGeneLocationBed}
\usage{
GenerateGeneLocationBed(genome_annotation, bedops_loc = NULL)
}
\arguments{
\item{genome_annotation}{Genome annotation DataFrame loaded with LoadGtf()
function in this package.}

\item{bedops_loc}{Optional. Location of BEDOPS in file system.}
}
\value{
Saves “gene_ranges.bed” in working directory.
}
\description{
Makes a bed file with gene boundaries, which is required for
assigning intergenic reads to a specific gene and discovering genes with large
amounts of intergenic reads near its 3’ gene end.

Note 1: This step is partially run in Linux Terminal in Bash and requires BEDOPS
(https://bedops.readthedocs.io/en/latest/). Make sure BEDOPS is installed and
provide a path to BEDOPS in the function if you get an error message.

Note 2: In Linux terminal, navigate to folder with the genome annotation of interest.
The annotation file should be named "genes.gtf" per 10x Genomics convention.
}
\examples{
genome_annotation <- LoadGtf(unoptimized_annotation_path = "test_genes.gtf")
GenerateGeneLocationBed(
genome_annotation = genome_annotation,
bedops_loc = NULL)
}
