% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyOverlappers.R
\name{IdentifyOverlappers}
\alias{IdentifyOverlappers}
\title{IdentifyOverlappers}
\usage{
IdentifyOverlappers(genome_annotation)
}
\arguments{
\item{genome_annotation}{Unoptimized genome annotation file in GTF. Could be
obtained from Ensembl, Refseq, 10x Genomics or elsewhere.}
}
\value{
Rank-ordered gene list of same-strand overlapping genes (“overlapping_gene_list.csv”).
}
\description{
Identifies all same-strand overlapping genes based on the unoptimized
genome annotation file in GTF, rank-orders them according to the number of gene
overlaps. Saves the list of overlapping genes in working directory as “overlapping_gene_list.csv”.
}
\examples{
genome_annotation <- LoadGtf(unoptimized_annotation_path = "test_genes.gtf")
IdentifyOverlappers(genome_annotation = genome_annotation)
}
