% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGtf.R
\name{LoadGtf}
\alias{LoadGtf}
\title{LoadGtf}
\usage{
LoadGtf(genes_gtf_path)
}
\arguments{
\item{genes_gtf_path}{Path to annotion .gtf file.}
}
\value{
Genome annotation file (genome_annotation).
}
\description{
Use to import the ENSEMBL/10x Genomics default genome annotation
file (GTF).
Note: This file can be downloaded from 10x Genomics provided reference
transcriptome "gene" folder at "https://support.10xgenomics.com/single-cell-gene-expression/software/downloads/latest"
or Ensembl.org if wish to customize more.
}
\examples{
LoadGtf("test_genes.gtf")
}
