% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizedAnnotationAssembler.R
\name{OptimizedAnnotationAssembler}
\alias{OptimizedAnnotationAssembler}
\title{OptimizedAnnotationAssembler}
\usage{
OptimizedAnnotationAssembler(
  exonic_gtf,
  gene_overlaps,
  gene_extension,
  gene_replacement
)
}
\arguments{
\item{exonic_gtf}{exonic gtf}

\item{gene_overlaps}{overlapping genes}

\item{gene_extension}{gene extension}

\item{gene_replacement}{gene replacement}
}
\value{
optimized annotation
}
\description{
OptimizedAnnotationAssembler
}
\examples{
genome_annotation <- LoadGtf("test_genes.gtf")
gene_overlaps <- IdentifyOverlappers(genome_annotation)
gene_extension <- GenerateExtensionCandidates()
OptimizedAnnotationAssembler(genome_annotation, gene_overlaps, gene_extension)
}
