% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapResolutions.R
\name{OverlapResolutions}
\alias{OverlapResolutions}
\title{OverlapResolutions}
\usage{
OverlapResolutions(genome_annotation, overlap_data, gene_pattern)
}
\arguments{
\item{genome_annotation}{ENSEMBL/10x Genomics default genome annotation file (GTF).}

\item{overlap_data}{A list of overlapping genes generated from IdentifyOverlappers.}
}
\value{
overlapping_gene_list.csv with added recommendations.
}
\description{
Based on original genome annotation .gtf file and a list of overlapping genes,
generates recommended actions for overlapping genes.
}
\examples{
genome_annotation <- LoadGtf("test_genes.gtf")
gene_overlaps <- IdentifyOverlappers(genome_annotation)
OverlapResolutions(genome_annotation, gene_overlaps, gene_pattern)
}
