% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PremrnaAnnotationGenerator.R
\name{PremrnaAnnotationGenerator}
\alias{PremrnaAnnotationGenerator}
\title{PremrnaAnnotationGenerator}
\usage{
PremrnaAnnotationGenerator(genome_annotation)
}
\arguments{
\item{genome_annotation}{ENSEMBL/10x Genomics default genome annotation file (.gtf).}
}
\value{
Generates a basic pre-mRNA reference and saves in working directory as premrna.gtf
}
\description{
It supplements original normal gene annotation entries by
traditional pre-mRNA entries where transcripts have been redefined as exons
and map in the --include-introns mode to retrieve most of available intronic reads.
}
\examples{
genome_annotation <- LoadGtf("test_genes.gtf")
PremrnaAnnotationGenerator(genome_annotation)
}
