% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateGeneLocationBed.R
\name{GenerateGeneLocationBed}
\alias{GenerateGeneLocationBed}
\title{GenerateGeneLocationBed}
\usage{
GenerateGeneLocationBed(genome_annotation, bedops_loc = NULL)
}
\arguments{
\item{genome_annotation}{Genome annotation file in .gtf format.}
}
\value{
Saves gene_ranges.bed in working directory.
}
\description{
Makes a bed file with gene boundaries, which is required for
assigning intergenic reads to a specific gene.

Note: This step is partially run in linux Terminal in Bash and requires
bedtools (https://bedtools.readthedocs.io/en/latest/content/installation.html).
Put bedtools in PATH after installing.

In linux terminal, navigate to folder with the genome annotation of interest.
Assuming that it is named "genes.gtf" per 10x Genomics convention.
}
\examples{
genome_annotation <- LoadGtf("test_genes.gtf")
GenerateGeneLocationBed(genome_annotation, bedops_loc)
}
