% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyOverlappers.R
\name{IdentifyOverlappers}
\alias{IdentifyOverlappers}
\title{IdentifyOverlappers}
\usage{
IdentifyOverlappers(genome_annotation)
}
\arguments{
\item{genome_annotation}{Genome annotation file in .gtf format.}
}
\value{
Rank-ordered gene list of same-strand overlapping genes (gene_overlaps).
}
\description{
Identifies all overlapping genes based on the ENSEMBL/10x Genomics
default genome annotation file (GTF), rank-order them according to the number of gene overlaps.
Prioritize this gene list for manual curation focusing on exonically overlapping genes.
Saves the list of overlapping genes in working directory (overlapping_gene_list.csv).
}
\examples{
genome_annotation <- LoadGtf("test_genes.gtf")
IdentifyOverlappers(genome_annotation)
}
